requirements.txt
slack_sdk>=3.26.1
python-dotenv>=1.0.1
slack_bolt>=1.28.1
